[VOID] Build: Restructure Phase 2
Segregated Tools/Editors into Toolbox, Docker, Timeline and ScriptEditor components

Toolbox is the home for Tools and Editors that provide additional functionalities for Void components
