+++ src/VoidDocker/Docker.cpp
@@ -17,9 +17,13 @@
 #include "VoidCore/Logging.h"
 #include "VoidIconForge/IconForge.h"
 #include "VoidQExtensions/Window.h"
-#include "VoidUi/Descriptors.h"
 
 VOID_NAMESPACE_OPEN
+
+namespace MimeTypes {
+    constexpr const char* TabIndex = "application-void/x-tab-index";
+} // namespace MimeTypes
+
 
 VoidDocker::VoidDocker(QWidget* parent)
     : QDockWidget(parent)
