+++ src/VoidScriptEditor/ScriptConsole.cpp
