+++ src/VoidScriptEditor/ScriptConsole.h
