+++ src/VoidScriptEditor/ScriptEditor.cpp
