+++ src/VoidScriptEditor/SyntaxHighlighter.cpp
