+++ src/VoidScriptEditor/SyntaxHighlighter.h
