+++ src/VoidTimeline/Timeline.cpp
@@ -15,7 +15,7 @@
 #include "VoidCore/Timekeeper.h"
 #include "VoidIconForge/IconForge.h"
 #include "VoidQExtensions/Tooltip.h"
-#include "VoidUi/Engine/Globals.h"
+// #include "VoidUi/Engine/Globals.h"
 
 static const int BUTTON_WIDTH = 30;
 static const int SMALL_BUTTON_WIDTH = 20;
@@ -197,7 +197,7 @@ void Timeline::Connect()
 
 void Timeline::Setup()
 {
-	SetFramerate(UIGlobals::FramerateString());
+	SetFramerate("24");
 	SetFrame(m_Timeslider->minimum());
 	m_PlayTimer.setTimerType(Qt::PreciseTimer);
