+++ src/VoidToolbox/Editor/Effects.h
