+++ src/VoidToolbox/Editor/Properties.cpp
