+++ src/VoidToolbox/Tools/Delegates/LogDelegate.cpp
