+++ src/VoidToolbox/Tools/Delegates/LogDelegate.h
