+++ src/VoidToolbox/Tools/LogWindow.cpp
